[IMP] web, *: move uuid in /web/core for frontend usage
There are multiple duplicates of the `uuid` function, presumably because it is
sometimes needed in frontend and was defined in `/web/views`.

This commit moves the function in `/web/core` for frontend availability and
removes duplicates.

task-4801746
