+++ addons/web/static/src/core/utils/strings.js
@@ -203,3 +203,15 @@ export function isNumeric(value) {
 export function exprToBoolean(str, trueIfEmpty = false) {
     return str ? !/^false|0$/i.test(str) : trueIfEmpty;
 }
+
+/**
+ * Generate a unique identifier (64 bits) in hexadecimal.
+ *
+ * @returns {string}
+ */
+export function uuid() {
+    const array = new Uint8Array(8);
+    window.crypto.getRandomValues(array);
+    // Uint8Array to hex
+    return [...array].map((b) => b.toString(16).padStart(2, "0")).join("");
+}
