+++ addons/web/static/src/views/fields/properties/properties_field.js
@@ -8,9 +8,8 @@ import { registry } from "@web/core/registry";
 import { user } from "@web/core/user";
 import { useBus, useService } from "@web/core/utils/hooks";
 import { useSortable } from "@web/core/utils/sortable_owl";
-import { exprToBoolean } from "@web/core/utils/strings";
+import { exprToBoolean, uuid } from "@web/core/utils/strings";
 import { useRecordObserver } from "@web/model/relational_model/utils";
-import { uuid } from "../../utils";
 import { standardFieldProps } from "../standard_field_props";
 import { PropertyDefinition } from "./property_definition";
 import { PropertyValue } from "./property_value";
